Update Source UD07

+++ kernel/include/arch/kernel_string.h
@@ -1,14 +1,15 @@
 #include <kernel.h>
 #include <stdint.h>
+#include <stddef.h>
 
-extern int strncmp(const char *strin, const char *strcmp, uint64_t n);
+extern int strncmp(const char *strin, const char *strcmp, size_t n);
 
-extern uint64_t strlen(const char *str);
+extern size_t strlen(const char *str);
 
 extern char *strrev(char *input, char *output);
 
-extern char *u64toh(uint64_t value, char *buffer);
+extern char *u64toh(size_t value, char *buffer);
 
-extern char *u64tos(uint64_t value, char *buffer);
+extern char *u64tos(size_t value, char *buffer);
 
 extern void kernel_printf(const char *format, ...);
