+++ kernel/src/arch/kernel_string.c
@@ -3,9 +3,10 @@
 #include <arch/kernel_string.h>
 #include <stdarg.h>
 #include <arch/serial.h>
+#include <stddef.h>
 
-int strncmp(const char *strin, const char *strcmp, uint64_t n){
-	for(uint64_t i = 0; i < n; i++){
+int strncmp(const char *strin, const char *strcmp, size_t n){
+	for(size_t i = 0; i < n; i++){
 		if(strin[i] != strcmp[i]){
 			return 0;
 		}
@@ -13,16 +14,16 @@ int strncmp(const char *strin, const char *strcmp, uint64_t n){
 	return 1;
 }
 
-uint64_t strlen(const char *str){
-	uint64_t ret = 0;
+size_t strlen(const char *str){
+	size_t ret = 0;
 	for(; *str; str++, ret++);
 	return ret;
 }
 
 char *strrev(char *input, char *output){
-	uint64_t len = strlen(input);
-	uint64_t i = 0;
-	uint64_t z = len;
+	size_t len = strlen(input);
+	size_t i = 0;
+	size_t z = len;
 	for(; i < len; i++){
 		output[i] = input[z - 1];
 		z--;
@@ -33,9 +34,9 @@ char *strrev(char *input, char *output){
 
 const char *ntoh = "0123456789abcdef";
 
-char *u64toh(uint64_t value, char *buffer){
+char *u64toh(size_t value, char *buffer){
 	static char buffer_n[17];
-	uint64_t i = 0;
+	size_t i = 0;
 	for(; value != 0; i++){
 		buffer_n[i] = ntoh[(value % 16)];
 		value /= 16;
@@ -49,9 +50,9 @@ char *u64toh(uint64_t value, char *buffer){
 	return buffer;
 }
 
-char *u64tos(uint64_t value, char *buffer){
+char *u64tos(size_t value, char *buffer){
 	static char buffer_n[21];
-	uint64_t i = 0;
+	size_t i = 0;
 	for(; value != 0; i++){
 		buffer_n[i] = ntoh[(value % 10)];
 		value /= 10;
@@ -68,8 +69,8 @@ char *u64tos(uint64_t value, char *buffer){
 void kernel_printf(const char *format, ...){ // note: works over serial, todo: implement working through the display
 	va_list args;
 	va_start(args, format);
-	uint64_t len = strlen(format);
-	for(uint64_t i = 0; i < len; i++){
+	size_t len = strlen(format);
+	for(size_t i = 0; i < len; i++){
 		switch(format[i]){
 			case '%':
 				i++;
@@ -90,12 +91,12 @@ void kernel_printf(const char *format, ...){ // note: works over serial, todo: i
 						serial_write_string(s);
 						break;
 					case 'u':
-						uint64_t u = va_arg(args, uint64_t);
+						size_t u = va_arg(args, size_t);
 						char buffer_u[21];
 						serial_write_string(u64tos(u, buffer_u));
 						break;
 					case 'x':
-						uint64_t x = va_arg(args, uint64_t);
+						size_t x = va_arg(args, size_t);
 						char buffer_x[17];
 						serial_write_string(u64toh(x, buffer_x));
 						break;
